# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.CreateClientRequest:
    properties:
      allowed_scopes:
        items:
          type: string
        type: array
      brand_color:
        type: string
      consent_message:
        type: string
      description:
        type: string
      is_confidential:
        type: boolean
      logo_url:
        description: UI設定フィールド
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      privacy_policy_url:
        type: string
      redirect_uris:
        items:
          type: string
        minItems: 1
        type: array
      require_consent:
        description: セキュリティ設定
        type: boolean
      support_email:
        type: string
      terms_of_service_url:
        type: string
      trusted_client:
        type: boolean
      website:
        type: string
    required:
    - name
    - redirect_uris
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@noraneko-id.com
    name: Noraneko ID Support
    url: https://noraneko-id.com/support
  description: プライベートサービス開発者向けIDaaS（Identity as a Service）API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://noraneko-id.com/terms
  title: Noraneko ID API
  version: "1.0"
paths:
  /admin/clients:
    post:
      consumes:
      - application/json
      description: 新しいOAuth2クライアントを作成します（管理者専用）
      parameters:
      - description: クライアント作成情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: クライアント作成成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: リクエストエラー
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 認証エラー
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 管理者権限が必要
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionAuth: []
      summary: OAuth2クライアント作成
      tags:
      - 管理者
...
securityDefinitions:
  BearerAuth:
    description: Bearer Token認証。「Bearer 」の後にアクセストークンを指定してください。
    in: header
    name: Authorization
    type: apiKey
  SessionAuth:
    description: セッション認証。session_tokenクッキーを使用します。
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"